Data and Ricker Models Updated
Up-to-date Nicola data (previous data in the "Old" folder within the "Data_in" folder now available and being referenced in the S-R code for Nicola. The Ricker models have been updated to use the new data, and now the SAS for BY 1992 is estimated in the "Ricker.stan" model (because of the mixobacteria issue that caused high mort in the hatchery smolts).