Simulation and Output Modification
Simulation code updated to test alpha across a controlled range of values according to quantiles of the Nicola alpha posterior from the SR model. Variation in the posterior for beta and sigma are preserved. The simulation is now run across 11 fixed values of alpha. The code to process the simulation output has been updated to handle multiple sets of simulation results