Lots of small changes
Simpars updated to reflect range of ERs we want to test. New lower benchmarks have been calculated to update the plots. Deleted the simulation results files as they are too heavy to commit and will be rerun with the updates to CUpars that Dan provided and the updates to the samSim package